move linear algebra files to specific directory and rename according to class
