+++ Common/include/linear_algebra/CSysVector.hpp
@@ -38,7 +38,7 @@
 
 #pragma once
 
-#include "./mpi_structure.hpp"
+#include "../mpi_structure.hpp"
 
 #include <climits>
 #include <cmath>
