+++ Common/src/linear_algebra/CSysSolve.cpp
@@ -35,8 +35,8 @@
  * License along with SU2. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "../include/linear_solvers_structure.hpp"
-#include "../include/linear_solvers_structure_b.hpp"
+#include "../../include/linear_algebra/CSysSolve.hpp"
+#include "../../include/linear_algebra/CSysSolve_b.hpp"
 
 template<class ScalarType>
 CSysSolve<ScalarType>::CSysSolve(const bool mesh_deform_mode) : cg_ready(false), bcg_ready(false),
