+++ SU2_CFD/include/solver_structure.hpp
@@ -62,9 +62,9 @@
 #include "../../Common/include/fem_geometry_structure.hpp"
 #include "../../Common/include/geometry_structure.hpp"
 #include "../../Common/include/config_structure.hpp"
-#include "../../Common/include/matrix_structure.hpp"
-#include "../../Common/include/vector_structure.hpp"
-#include "../../Common/include/linear_solvers_structure.hpp"
+#include "../../Common/include/linear_algebra/CSysMatrix.hpp"
+#include "../../Common/include/linear_algebra/CSysVector.hpp"
+#include "../../Common/include/linear_algebra/CSysSolve.hpp"
 #include "../../Common/include/grid_movement_structure.hpp"
 #include "../../Common/include/blas_structure.hpp"
 #include "../../Common/include/graph_coloring_structure.hpp"
